Should fix the slow rockets bugs.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@2449 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/server/progdefs.h
@@ -90,7 +90,7 @@ typedef struct entvars_s
 	vec3_t	absmin;
 	vec3_t	absmax;
 	float	ltime;
-	float	lastruntime;
+	int		lastruntime;
 	float	movetype;
 	float	solid;
 	vec3_t	origin;
